backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# File Storage & AWS
boto3==1.34.0
botocore==1.34.0

# Date/Time
python-dateutil==2.8.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
click==8.1.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0

# CORS
fastapi-cors==0.0.6
# python-cors==1.7.0

# Rate Limiting
slowapi==0.1.9

# Pagination
fastapi-pagination==0.12.13

# File Upload
python-magic==0.4.27

# PDF Generation (for reports)
reportlab==4.0.7

# Excel Export
openpyxl==3.1.2

# CSV Handling
pandas==2.1.4

# Background task monitoring
flower==2.0.1

# Database connection pooling
sqlalchemy-utils==0.41.1

# Timezone handling
pytz==2023.3

# Validation
validators==0.22.0

# Caching
fastapi-cache2==0.2.1

# API Documentation
fastapi==0.104.1
